internal: move type Version to version.go
