+++ internal/version_test.go
@@ -1,0 +1,32 @@
+package internal
+
+import "testing"
+
+func TestVersion(t *testing.T) {
+	a, err := NewVersion("1.2")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	b, err := NewVersion("2.2.1")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if !a.Less(b) {
+		t.Error("A should be less than B")
+	}
+
+	if b.Less(a) {
+		t.Error("B shouldn't be less than A")
+	}
+
+	v200 := Version{2, 0, 0}
+	if !a.Less(v200) {
+		t.Error("1.2.1 should not be less than 2.0.0")
+	}
+
+	if v200.Less(a) {
+		t.Error("2.0.0 should not be less than 1.2.1")
+	}
+}
